requirements.txt
pypdf==3.17.4
reportlab==4.0.8
Pillow==10.2.0